Conflicts:
	Fish.MovieManager.Test/Program.cs
	Model/Fish.MovieManager.DoubanAPI/Class.cs
	Model/Fish.MovieManager.DoubanAPI/Fish.MovieManager.DoubanAPI.csproj
	MyModernMovieManager.sln

+++ Model/Fish.MovieManager.VideoFileInfo/Storage/VideoFileInfo.cs
@@ -63,7 +63,7 @@ namespace Fish.MovieManager.VideoFileInfo.Storage
         /// <summary>
         /// 视频总帧数
         /// </summary>
-        public virtual int totalFrames { get; set; }
+        public virtual long totalFrames { get; set; }
         /// <summary>
         /// 用户的评分
         /// </summary>
